% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{default_sim_seed_42}
\alias{default_sim_seed_42}
\title{Example simulation from running `salmon_sim()` with some specified (not
default) values, before beta values were changed to be proportions and sum
to 1. See tests for compring this and `default_sim_seed_42_new`.}
\format{
Tibble - see `?salmon_sim()`.
}
\source{
Generated from running code before the betas were changed to sum to 1.
}
\usage{
default_sim_seed_42
}
\description{
Example simulation from running `salmon_sim()` with some specified (not
default) values, before beta values were changed to be proportions and sum
to 1. See tests for compring this and `default_sim_seed_42_new`.
}
\keyword{datasets}
