% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{default_sim_seed_42_new}
\alias{default_sim_seed_42_new}
\title{Example simulation from running `salmon_sim()` with some specified (not
default) values.}
\format{
Tibble - see `?salmon_sim()`.
}
\source{
Generated from running `data-raw/default_sim_seed_42_new.R`.
}
\usage{
default_sim_seed_42_new
}
\description{
Example simulation from running `salmon_sim()` with some specified (not
default) values.
}
\keyword{datasets}
