% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{default_sim_seed_42_new_fit}
\alias{default_sim_seed_42_new_fit}
\title{Example simulation and fit from running `salmon_sim()` with some specified (not
default) values and then `pbsEDM::pbsEDM()`, to test with use of `sim_and_fit()`.}
\format{
List of fitting results - see `?pbsEDM::pbsEDM()`.
}
\source{
Generated from running `data-raw/default_sim_seed_42_new.R`.
}
\usage{
default_sim_seed_42_new_fit
}
\description{
Example simulation and fit from running `salmon_sim()` with some specified (not
default) values and then `pbsEDM::pbsEDM()`, to test with use of `sim_and_fit()`.
}
\keyword{datasets}
