% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{fit_models}
\alias{fit_models}
\title{Fit EDM, Larkin, and Ricker models}
\usage{
fit_models(
  all_sims,
  res_realisations,
  R_switch,
  T,
  pbsEDM_args,
  larkin_args,
  ricker_args
)
}
\arguments{
\item{all_sims}{List of simulated data generated from salmon_sim()}

\item{res_realisations}{Empty data frame with correct headers for outputs 
of estimation}

\item{R_switch}{either `R_prime_t` or `R_t` to specify which one
the calculations are based on.}

\item{T}{to be filled in}

\item{pbsEDM_args}{List of arguments to pass onto `pbsEDM::pbsEDM()`. Note
that `lags` has no default, so needs to be specified here. First one
must be called `R_switch` to then get automatically changed to what is
specified by the `R_switch` parameter}

\item{larkin_args}{List of arguments to pass onto `larkin::....()`.}

\item{ricker_args}{List of arguments to pass onto `larkin::....()`.}
}
\value{
List object call output with components:
 - `single_realisation` forecasts from all models in last year
 - `edm_fit_single` predictions for each year from EDM
 - `edm_lar_single` predictions for each year from larkin
 - `edm_ric_single` predictions for each year from ricker
}
\description{
Fits EDM, Larkin, and Ricker models to simulated data using 
  parallel processing
}
\examples{
\dontrun{
}
}
\author{
Carrie Holt and Andrew Edwards
}
