% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_errors.R
\name{forecast_errors}
\alias{forecast_errors}
\title{Calculate forecasting errors}
\usage{
forecast_errors(run, label)
}
\arguments{
\item{run}{output from fit_models() for EDM, Larkin and Ricker models}

\item{label}{name of sensitivity run for plotting purposes}
}
\value{
List object call output with components:
 - `forecasts` forecast for the last year, from all models for each 
 realisation 
 - `absErr` absolute error in forecasts in the last year, from all models 
 for each realisation
}
\description{
Calculate forecasting errors
}
\examples{
\dontrun{
}
}
\author{
Carrie Holt and Andrew Edwards
}
