% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salmon_bif.R
\name{plot_salmon_bif}
\alias{plot_salmon_bif}
\title{Plot a bifurcation-type diagram}
\usage{
plot_salmon_bif(alpha_vec, x, col = "black", ...)
}
\arguments{
\item{alpha_vec}{Vector of alpha values to try}

\item{x}{Output from salmon_bif()}
}
\value{
Figure gets plotted
}
\description{
Plot the last values of recruitment against alpha.
}
