% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salmon_sim.R
\name{plot_sim}
\alias{plot_sim}
\title{Plot output from salmon_sim()}
\usage{
plot_sim(x, new_plot = TRUE, ...)
}
\arguments{
\item{x}{Output data frame from salmon_sim().}

\item{new_plot}{Start a new plot or not.}
}
\value{
Plot of simulated spawners against time.
}
\description{
Plot output from salmon_sim()
}
