% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotLag2d}
\alias{plotLag2d}
\title{Create 2-d pdf movie of lagged points}
\usage{
plotLag2d(x, pdf_filename, start = 1, end = length(x),
  late_col = "red", early_col = "black",
  early_col_lines = "lightgrey", late_num = 10, pt_type = "p",
  x_lab = expression("x"[t - 1]), y_lab = expression("x"[t]),
  figheight = figheight, figwidth = figwidth, ...)
}
\arguments{
\item{x}{Vector of values}

\item{pdf_filename}{Filename (including .pdf) to save the movie}

\item{start}{First value to use when plotting}

\item{end}{Final value to use when plotting}

\item{late_col}{Colour for the late points (the last 'late_num' points) for
each frame}

\item{early_col}{Colour for the early points}

\item{early_col_lines}{Colour for the early lines}

\item{late_num}{How many points to have colour 'late_col'}

\item{pt_type}{Type of point (the usual plot(..., type= ) option, but do not
use "o".}

\item{x_lab}{Label for x axis}

\item{y_lab}{Label for y axis}

\item{figheight}{Figure height}

\item{figwidth}{Figure width}
}
\value{
Movie as a .pdf then can then be easily input into Sweave file using
  the animategraphics latex pacakge. [Update this regarding after testing knitr etc.]
}
\description{
Create 2-d movie of x_t (on the y-axis) against x_{t-1}, with colours to
show the latest points as it grows.
}
\examples{
\dontrun{
  plotLag2d(x = default_sim_seed_42$p_t3, pdf_filename = "hello.pdf")  # not
  a time series
}
}
\author{
Andrew Edwards
}
