% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salmon_bif.R
\name{salmon_bif}
\alias{salmon_bif}
\title{Create calculations for bifurcation-type plot for deterministic simulations}
\usage{
salmon_bif(alpha_vec = seq(0.01, 10, by = 0.01), last = 50,
  T = 10000, new_plot = TRUE, ...)
}
\arguments{
\item{alpha_vec}{Vector of alpha values to use.}

\item{last}{How many of the final time steps to save.}

\item{T}{Number of years to run the simulation.}

\item{new_plot}{Start a new plot or not.}

\item{...}{Further inputs for salmon_sim().}
}
\value{
matrix of final R_t values as columns, with each column
  corresponding to a value of alpha
}
\description{
Calculates the last values for a deterministic run for different values of (for
  now) alpha.
}
