% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salmon_sim.R
\name{salmon_run}
\alias{salmon_run}
\title{Simulate Larkin model and plot results}
\usage{
salmon_run(..., new_plot = TRUE)
}
\arguments{
\item{...}{Inputs for salmon_sim().}

\item{new_plot}{Start a new plot or not.}
}
\value{
Plot of results.
}
\description{
Simulate Larkin model and plot results
}
