% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{shift}
\alias{shift}
\title{Function that shifts vector values to right or left}
\usage{
shift(x, n, invert = FALSE, default = NA)
}
\arguments{
\item{x}{Vector for which to shift values}

\item{n}{Number of places to be shifted.
Positive numbers will shift to the right by default.
Negative numbers will shift to the left by default.
The direction can be inverted by the invert parameter.}

\item{invert}{Whether or not the default shift directions
should be inverted.}

\item{default}{The value that should be inserted by default.}
}
\description{
Taken from https://stackoverflow.com/questions/26997586/shifting-a-vector by
  user petermeissner.
}
