% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simple_ts}
\alias{simple_ts}
\title{Results of simple spawner simulation that shows Issue #15 in rEDM}
\format{A vector containing 99 first-differenced values of spawners from a simple
simulation. Using it here as it shows Issue #15 in rEDM.
It is originally from sockeye-sim-edm.rnw that used Carrie's original
function:
 set.seed(42)
 set.seed(42)
 simulated = salmonTraj() # Simulated annual spawner abundances and recruitments
 N = simulated$S             # Just the simulated spawners
 T = length(N)               # Corresponding years
 tvec = 1:T
 x = N[-1] - N[-length(N)]   # x = first-differences
 simple_ts <- x}
\usage{
simple_ts
}
\description{
Results of simple spawner simulation that shows Issue #15 in rEDM
}
\keyword{datasets}
