% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{three_sim_fits}
\alias{three_sim_fits}
\title{Example simulations and fits from running `sim_and_fit_realisations()`.}
\format{
List of sim_and_fit_realisations fitting results - see `?sim_and_fit_realisations`
}
\source{
Generated from running `data-raw/three_sim_fits.R`
}
\usage{
three_sim_fits
}
\description{
Example simulations and fits from running `sim_and_fit_realisations()`.
}
\keyword{datasets}
