% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{default_sim_seed_42}
\alias{default_sim_seed_42}
\title{Results of salmon_sim() with original default inputs, and set.seed(42)}
\format{A data frame with 100 rows and 11 columns, obtained from running
  set.seed(42)
  salmon_sim()
  usethis::use_data(default_sim_seed_42, overwrite=TRUE)
  using early defaults of the inputs (which are specified in the test).}
\usage{
default_sim_seed_42
}
\description{
A dataframe containing the results of the original simulation.
}
\keyword{datasets}
