% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salmon_sim.R
\name{salmon_run}
\alias{salmon_run}
\title{Simulate Larkin model and plot results}
\usage{
salmon_run(..., new_plot = TRUE)
}
\arguments{
\item{...}{Inputs for salmon_sim().}

\item{new_plot}{Start a new plot or not.}
}
\value{
Plot of results, and matrix of results from salmon_sim().
}
\description{
Simulate Larkin model and plot results
}
\examples{
\dontrun{
  salmon_run(alpha = 1.5,
             beta = c(0, 0.8, 0, 0),
             p_prime=c(0, 1, 0),
             T = 1000,
             deterministic = TRUE)# no densitity dependence on alternate
                                  #  years, so unconstrained increase in
                                  #  population
}
}
