% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salmon_sim.R
\name{salmon_sim}
\alias{salmon_sim}
\title{Simulation of Larkin population dynamics model}
\usage{
salmon_sim(alpha = 0.8, beta = c(0.8, 0.2, 0.1, 0.1),
  p_prime = c(0.01, 0.98, 0.01), rho = 0.6, omega = 0.8,
  sigma_nu = 0.75, phi_1 = 0.1, T = 100, h_t = NULL,
  R_t_init = c(0.6, 0.1, 0.1, 0.1, 0.6, 0.1, 0.1, 0.1),
  deterministic = FALSE, extirp = 2e-06)
}
\arguments{
\item{alpha}{Ratio of recruits to spawners at low spawner abundance in the
absence of noise.}

\item{beta}{Vector [beta_0, beta_1, beta_2, beta_3] that scales
the magnitude of density dependence based on the current spawning stock
(beta_0) and the previous three years (beta_1, beta_2 and beta_3).
beta_1 = beta_2 = beta_3 = 0 reduces the model to a Ricker model.}

\item{p_prime}{Vector of typical proportion of recruits spawned in a year that will
come back to freshwater as age-3, age-4 and age-5 (each of three elements
in the vector, which must sum to 1).}

\item{rho}{Autocorrelation parameter for process noise, >= 0.}

\item{omega}{Scales the annual normal deviates on the proportions returning
at each age. If omega = 0 then no stochasticity for the proportions.}

\item{sigma_nu}{Standard deviation of process noise. If sigma_nu = 0 and
rho = 0 then no process noise.}

\item{phi_1}{Initial value of process noise.}

\item{T}{Number of years to run the simulation, including the eight years
needed for initialising the simulation, must be >=9.}

\item{h_t}{Vector of harvest rate for each year 1, 2, 3, ..., T. Or if a
single value then this will be the constant rate for all years. If NULL
then harvest rate will be set to 0.2 for all years.}

\item{R_t_init}{Vector of eight years of recruit abundance to initialize the
model.}

\item{deterministic}{If TRUE then include no stochasticity and any
given values of rho, omega, sigma_nu and phi_1 are redundant and set to
0.}

\item{extirp}{Value below which we consider the population extirpated, in
same units as recruits and spawners.}
}
\value{
Matrix of years (rows) with named columns:
  t: year;
  R_t: total recruits returning in year t;
  R_prime_t: number of adult recruits generated from spawners in year t that
    will return to freshwater (and then be subject to fishing and can then
    spawn);
  S_t: number of fish that spawn in year t;
  h_t: harvest rate in year t;
  p_t3, p_t4, p_t5: proportion of R_prime in year t that later returned at age 3,
    4 and 5.
}
\description{
Simulates Larkin population dynamics model, based on Appendix F of Holt
  et al. 2018/011 Res Doc (and what's in our draft manuscript), Ricker
  Appendix, and Grant 2010/042 Res. Doc.
  Notation matches that of our write up.
  The harvest rate has to be prescribed for all years, and to initialize the
  model the recruitment R_t has to be given for the first eight years. All
  parameters must be >=0, some >0. Units of recruits and spawners are
  millions of fish, and hence beta has units 1/(millions of fish).
}
