% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_and_fit.R
\name{sim_and_fit}
\alias{sim_and_fit}
\title{Simulate a population and fit it using pbsEDM::pbsEDM(), for getting
 full results for any specific simulation.}
\usage{
sim_and_fit(salmon_sim_args = list(), pbsEDM_args = list())
}
\arguments{
\item{salmon_sim_args}{List of arguments to pass onto `salmon_sim()`.}

\item{pbsEDM_args}{List of arguments to pass onto `pbsEDM::pbsEDM()`. Note
that `lags` has no default, so needs to be specified here.}
}
\value{
List object with components:
 - `simulated:` the simulated tibble
 - `fit:` the full output from pbsEDM::pbsEDM().
}
\description{
Simulate a population and fit it using pbsEDM::pbsEDM(), for getting
 full results for any specific simulation.
}
\examples{
\dontrun{
res <- sim_and_fit(pbsEDM_args = list(lags = list(R_prime_t = 0,
                                                  S_t = 0:3),
                                      first_difference = TRUE))
res$fit$results$X_rho
}
}
\author{
Andrew Edwards
}
